+++ src/app/navigation/navigation.module.ts
@@ -77,6 +77,7 @@ import { AutofillDirective } from '../autofill.directive';
 import { ShortcutButtonStyleComponent } from '../styles/shortcut-button-style/shortcut-button-style.component';
 import { OpenAppButtonStyleComponent } from '../styles/open-app-button-style/open-app-button-style.component';
 import { CheckboxStyleComponent } from '../styles/checkbox-style/checkbox-style.component';
+import { LabJSComponent } from '../styles/lab-js/lab-js.component';
 PlotlyModule.plotlyjs = PlotlyJS;
 
 const routes: Routes = [
@@ -186,7 +187,8 @@ const routes: Routes = [
         AutofillDirective,
         ShortcutButtonStyleComponent,
         OpenAppButtonStyleComponent,
-        CheckboxStyleComponent
+        CheckboxStyleComponent,
+        LabJSComponent
     ]
 })
 export class NavigationPageModule {