+++ src/app/styles/lab-js/lab-js.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-lab-js',
+  templateUrl: './lab-js.component.html',
+  styleUrls: ['./lab-js.component.scss'],
+})
+export class LabJSComponent  implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {}
+
+}